[최현준] 운영자 테스트 후 페이징, 삭제버튼, textbook 추가, 승인대기상태 변경
테스트 후 페이징 오류가 발생한 코드 일괄 수정, 삭제버튼 클릭 시 경고창 발생 후 확인버튼 클릭했을때 삭제기능 추가,
textbook 추가 시 고유번호 중복 검사 추가, 승인대기 상태인 강사/학생이 없을 시 메세지 출력기능 추가

+++ lms/src/main/webapp/WEB-INF/view/manager/classroomList.jsp
@@ -71,7 +71,7 @@
 													<td>${c.classroomTotal }</td>
 													<td>${c.classroomState }</td>
 													<c:if test="${c.classroomState =='빈강의실'}">
-														<td><a class="btn btn-outline-danger" style="border-radius: 4px;"href="${path }/manager/deleteClassroom/${c.classroomNo}/${currentPage}">삭제</a></td>
+														<td><a onclick="return confirm('정말로 삭제하시겠습니까?');" class="btn btn-outline-danger" style="border-radius: 4px;"href="${path }/manager/deleteClassroom/${c.classroomNo}/${currentPage}">삭제</a></td>
 													</c:if>
 													<c:if test="${c.classroomState =='사용중'}">
 														<td>&nbsp;</td>
